test(huddle): exercise STT worker floor cleanup
